+++ ntpd/bin/ntp-daemon.rs
@@ -1,6 +1,9 @@
 #![forbid(unsafe_code)]
 
+use std::process;
+
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    ntpd::daemon_main().await
+async fn main() {
+    let result = ntpd::daemon_main().await;
+    process::exit(if result.is_ok() { 0 } else { 1 });
 }